test: 🧪 add e2e folder
