+++ package.json
@@ -6,7 +6,8 @@
 		"dev": "doppler run -- vite dev --open",
 		"build": "doppler run -- vite build",
 		"preview": "doppler run -- vite preview",
-		"test": "playwright test --repeat-each=4",
+		"test": "playwright test",
+		"test:gen": "playwright codegen localhost:5173",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --plugin-search-dir . --check . && eslint .",
