+++ src/routes/(e2e)/cookie-banner.e2e.ts
@@ -1,0 +1,14 @@
+import test, { expect } from '@playwright/test';
+
+test('dismisses cookie banner', async ({ page }) => {
+	await page.goto('/', { waitUntil: 'networkidle' });
+	const banner = page.getByTestId('cookie-banner');
+
+	await expect(banner).toBeVisible();
+	await expect(banner).toHaveCount(1);
+
+	await page.getByTestId('cookie-banner-btn').click();
+
+	// element does not exist in the dom
+	await expect(banner).toHaveCount(0);
+});
